third_party/blink/web_tests/fast/arraybuffer/size_check_of_web_api.html

<!DOCTYPE html>
<html crossorigin>
<head>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
</head>
<body>
  <script>
test(() => {
  let mem = new WebAssembly.Memory({initial: 33000, maximum: 33000});
  let canvas = document.createElement("canvas");
  let context = canvas.getContext("webgl");
  let data = new Uint8Array(mem.buffer);
  assert_throws_js(RangeError,
    () => context.bufferSubData(context.ARRAY_BUFFER, 0, data));
}, "Test that webgl rejects an ArrayBuffer bigger than 2GB.");

test(() => {
  let mem = new WebAssembly.Memory({initial: 33000, maximum: 33000});
  let canvas = document.createElement("canvas");
  let context = canvas.getContext("webgl");
  let data = new Uint8Array(mem.buffer, 10, 16);
  context.bufferSubData(context.ARRAY_BUFFER, 0, data);
}, "Test that webgl accepts a small view of ArrayBuffer bigger than 2GB.");
  </script>
</body>
</html>
